[FIX] account_accountant: duplicate rpc in BankRecKanban
Since the new relational model odoo/odoo#114024, in the kanban reconciliation view,
when the card is switched, two similar "web_search_read" RPCs are executed.

Why:
Because two renderings are caused, one because of the update applied to
the record datapoint and the second when the newState is applied to the kanban view.

Solution:
Customise the Relational Model to have a custom update for "todo_command".
The changes will be applied when the newState is applied so that there is
only one rendering.

closes odoo/enterprise#47199

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>