[FIX] l10n_be_intrastat: export reports conditionally
Depending on the filters applied on the intrastat report, different
reports should be added to the XML export. Different companies have
different reporting obligations when it comes to intrastat reports in
Belgium. This largely depends on the total value of goods exchanged, but
is specified by the government, and is largely unknown to us.

Only the reports relevant to a particular company should be included in
the XML export. Including a report for which the company has no
obligation results in an error while uploading.

Our way of solving this is by having the XML contain only the reports
which the current filters allow. The relevant filters in this case are
"arrivals" and "dispatches", and "extended" (vs "standard").

Tests have been added to ensure the export format of the
l10n_be_intrastat report is correct.

closes odoo/enterprise#47773

Task-id: 3501515
X-original-commit: 7cbf8ccda0434cfad150b8e6ffa96c1596111ac5
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>