[IMP] documents_spreadsheet: create spreadsheet from template with user locale
Currently, when creating a spreadsheet from a template,  numbers are always
displayed with the en_US format, no matter the user's lang.

That's because the locale is hardcoded in the spreadsheet source file, in
the source code (or it falls back to the default en_US locale if it's missing).

With this commit, we change the spreadsheet locale with the user's locale when
the spreadsheet is created.

Limitation
----------
Since hardcoded date formats are not changed when the locale changes, dates
will keep the en_US format. However, most (if not all) dates in templates
are coming from ODOO.PIVOT and ODOO.LIST functions, which compute the formats
dynamically based on the current locale.

Task: 3484002
Part-of: odoo/enterprise#47284