[IMP] account_accountant: Better UX during account_accoutant_tour
Description of the issue/feature this commit addresses:

The current onboarding account tour can be enhanced.
The purpose of this commit is to improve it.

Desired behavior after the commit is merged :

This commit modifies the onboarding account tour to clarify the messages that
are shown, add messages to certain steps and try to make it impossible
for the user to be lost during its entierty.

Adding this improvement, the onboarding account tour better fills
its purpose.

task-3375344

closes odoo/enterprise#47306

Related: odoo/odoo#135187
Signed-off-by: Quentin De Paoli <qdp@odoo.com>