[FIX] account_reports: Dispatch actions with the right options
Before this commit, the fact the export buttons were resolved on the sections source of the report was enforced on the js side, while on server-side, the test calling the appropriate function had to explicitly mimic that.

This caused confusion, to the point we did not realize that, when using sections, export buttons where resolved on the section source, but used the rerouted options of the opened section for that. This could create errors at exports, if the section did not support the exact same filters as the composite report owning it ; in turn, depending on the code related to the button, this could cause issues.

We remove that variability by ensuring the options used to dispatch a function are compatible with the report. In case of a composite report, we regenerate the options. To do that, we also deconfuse the way the sections source is selected to dispatch button functions, and tweak a little bit the computation of the 'report_id' option key so that it does not reroute anymore in this case, for clarity.

closes odoo/enterprise#47380

X-original-commit: 858d98488acf2148b69c21b1429e89ee6a7672be
Signed-off-by: John Laterre (jol) <jol@odoo.com>