[FIX] documents_spreadsheet: do not crash when inserting link
This commit comes with odoo/odoo#135300, which fixes an issue with
the "Insert menu in spreadsheet" feature. In stock.picking, the
domain of the action is the empty string. In this case, it crashed
because the empty string isn't a valid domain representation.

This commit adds the test case.

closes odoo/enterprise#47363

X-original-commit: 81d7508405d35772672626571eaeed562f6f37b0
Related: odoo/odoo#135369
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>