[FIX] documents_spreadsheet: do not crash when inserting link
This commit comes with odoo/odoo#135300, which fixes an issue with
the "Insert menu in spreadsheet" feature. In stock.picking, the
domain of the action is the empty string. In this case, it crashed
because the empty string isn't a valid domain representation.

This commit adds the test case.

closes odoo/enterprise#47500

X-original-commit: 9cf4277aa057929ea9b39fc862e2e73eaaf7983a
Related: odoo/odoo#135738
Signed-off-by: Samuel Degueldre (sad) <sad@odoo.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>