[FIX] whatsapp: neutralize
This commit adds the missing neutralisation necessary for the `whatsapp`
module introduced in [1]

The purpose of the standard neutralisation framework is to allow us to
create database copies that will not interract with external systems in
ways that could impact the production database (or if it is not possible
to prevent the interractions, make sure that they are benign or wont result in actual changes),
or impact the customers of the operator of the production database.

This is mainly useful to allow safe support investigation on database
duplicates.

[1] https://github.com/odoo/enterprise/pull/45601

closes odoo/enterprise#49338

X-original-commit: 6e7078d0c80165cfca2cc2801263a6c342bac717
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Alexandre Moens (mao) <mao@odoo.com>