[FIX] account_reports: multicurrency revaluation report action
It was not possible to open the General Ledger from the multicurrency
revaluation because of a Traceback.

There were 2 problems:

The 'reportAction' was expecting an 'ev' parameter but since the
action was added in a , there was none. So there was a traceback
when trying to do a preventDefault/stopPropagation on the 'ev'. To fix
it, we just check if 'ev' is defined before.

When calling the action, there was another traceback because we were
doing a 'self.ensure_one()' but there was no record on 'self' which
makes sense since we are in a report handler. To fix it, we have to get
the report manually with a browse.

closes odoo/enterprise#47391

X-original-commit: df4f047f0440e2b3e2dc04ae884a6e02ccd6333f
Signed-off-by: John Laterre (jol) <jol@odoo.com>