[IMP] appointment: add videocall link preference
Purpose
=======
Give full control to users on which link will be used for the
events Odoo will generate when appointments are booked.

Specifications
==============
Removing the correlation between the location and the videocall
link generation by adding a selection field called
"Videoconference Link" in the appointment form view.
This field will give full control to the user to specify which
videocall link source to give to their clients when an appointment
from this appointment type is booked.
The default videocall link source is the Odoo Discuss one. The user could
also choose to not give any link at all.

Task-32856011

Part-of: odoo/enterprise#46688