[FIX] account_{consolidation, reports}: edit popover
The boolean edit popover was not working properly. The python was
expecting a boolean in the form of an integer (0 or 1) but the
javascript was sending an actual boolean (true or false). It was fixed
by just sending the boolean converted to a number.

The styling of the boolean popover was missing, so it was added in this
commit as well.

closes odoo/enterprise#47594

X-original-commit: b05aedab5f549715be614db36fdd762bdac8f494
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Ruben Gomes (rugo) <rugo@odoo.com>