[FIX] documents_project: show documents settings in project form view
In the view refactor of https://github.com/odoo/enterprise/pull/35255, the condition to show
the rating and documents section in the settings of the project form view was
converted to `True` instead of `False`, meaning that the rating and documents settings
are never visible with documents installed.

This commit fixes the visibility so the section is always visible with documents installed instead.

closes odoo/enterprise#47392

Signed-off-by: Audric Onockx (auon) <auon@odoo.com>