[FIX] account_reports: error when reloading PL
When opening the partner ledger for a customer
from the customer page, an error is raised when
reloading the report, this is because the report
id is take from the action context, which is reset
when reloading. With this commit, if there is no
report id in the action context, we take it from
the url parameters.

Steps:

- Go to Accounting > Customers > Customers
- Open a customer form view and open his
  Partner Ledger
- Reload the page
-> Error: "Cannot read properties of undefined
   (reading 'report_id')"

opw-3461791

closes odoo/enterprise#47398

Signed-off-by: John Laterre (jol) <jol@odoo.com>