[FIX] web_studio: ReportEditor: qweb context node within a `t-call`
To reproduce
============
- on any app open studio
- create blank report
- try to add a field => client Error

Problem
=======
when trying to add a field, we look for the node containing the context,
in this case it's not found which leads to the Error

This was because of two things.
- first, the t-calls were not correctly inlined: the `<t t-set="..." />` were wrapped inside
the value used to inline the `<t t-raw="0" />` expressions down the line when a HTML body was present inside the t-call.

- second, because all the qweb is not relevant to the ReportEditor (e.g. the html_container), the QWeb nodes outside of the `div#wrapwrap` were not present as such in the reportEditor, hence, any variable defined outside the #wrapwrap node was not defined from the reportEditor point of view

Solution
========
The inlining of t-calls has been corrected. All `t-sets` that are direct children of the t-call remain unmoved.
The body, if it exists, is moved to the variable used to transmit to the `t-raw="0"` nodes down the line

The current context of the report's inlined Qweb is applied on the #wrapwrap node, which represents the relevant and visible part of a report, and the main editable element.

opw-3499802

closes odoo/enterprise#48883

X-original-commit: da7a7a0f59f6cadd9ec3aa9c0d5c66b26a84f203
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>