[FIX] l10n_be_soda: missing names imported lines
Since saas-16.4, the field `name` is required
on the new model `soda.account.mapping`, which is
wrong because a coda file could be imported
without label on transaction line. Therefore, a
ValidationError is raised in this case.

With this commit, we set the label to an empty
string if label is not set. We also adapt a test
and the related test file to be consistent with
the behavior.

opw-3506519

closes odoo/enterprise#48821

X-original-commit: 362238cf2683c764d693af496b3fd5a8535da376
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>