[FIX] documents_project: avoid copying workspaces when copying projects
Steps to reproduce
---

1. Install Project and Documents.
2. Create a project with the documents feature enabled and take note of the workspace name.
3. Duplicate the project.
4. Go to Documents > Workspace, you will see that there are two copies of the original workspace.

---

This occurs because, although the project_workspace is set as `copy=False`, the `copy` methods calls the
`create` method, in which we automatically create folders for new projects unless there is the context
key `no_create_folder=False`.

This commit fixes the issue by temporarily adding that context key to the context, until
the call to create is made.

Task-3491191

closes odoo/enterprise#47457

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>