[IMP] knowledge: add new action buttons in tree view
This commit adds new action buttons to the Knowledge's tree view when selecting
articles. We added the buttons `Restore` and `Send to Trash`.

`Restore` is seen in the Trash view and `Send to Trash` in the Search list view.
Those are used to either send an article to the trash or to restore them from
it in batch.

task-3470267

closes odoo/enterprise#47502

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>