[IMP] account_reports: control panel buttons
Creates a new component 'AccountReportButtonsBar' that allows to define
two types of control panel buttons: 'single' and 'grouped'.

'Grouped' buttons will be display in a dropdown and 'single' buttons
will always be displayed separately.

To define a 'single' button, the key 'always_show: True' should be added
at the button definition in options['buttons'] in the account report.

The default behavior is 'grouped'.

closes odoo/enterprise#47514

X-original-commit: c550fe7e0955c84479e25574c811205b77e4bf0b
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>