[FIX] knowledge: fixes link sent via mail for portal users
This commit fixes a bug that impacted the invite links sent via mail to portal
users. The links sent redirected the user to the home page instead of the
article on which he/she was invited.

This bug was caused by a call to a server action that didn't use the correct
context key to reference the article.

Now the server action's context uses the correct keys and redirects to the
correct article.

task-3505408

closes odoo/enterprise#47626

X-original-commit: ba6ed51ea296297e3d200101da7de11c22eb4ecf
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Thomas Josse (thjo) <thjo@odoo.com>