[FIX] partner_commission: add chatter message if draft a refund
Steps to reproduce:
-------------------
- create a commission plan (e.g. 10% on everything);
- assign this commission plan to a partner level;
- create a customer;
- create a partner and add the partner level;

- create a product;
- create an SO with the created product and the created partner as the referrer;
- create and validate the invoice;
- register the payment which will create a RFQ for the commission of 10%;

- go back to the invoice and create a credit note and confirm it;
- on the invoice, unreconcile the journal entries;
- add the outstanding credit of the reversal;

- checking the RFQ, an order line was added due to the created credit note (RFQ is now 0);

- go back to the credit note and reset it to draft to edit parts of it;
- edit part (or not);
- confirm it and add the outstanding debit to the related invoice;

Issue:
------
A second negative order line as added to the RFQ.
The amount on the RFQ is negative.

Cause:
------
Validating a refund on an invoice with commission creates
a purchase order line with the refund data.

Unfortunately, put this refund in draft does not update the sale order.
And when we re-confirm the sale order,
we will create a new purchase order line.

Solution:
---------
Add a message in the chatter to warn the user and keep track
of refunds that may be duplicated.

opw-3377288

closes odoo/enterprise#48079

X-original-commit: 409c3c82acd467119a8f82cbc8d1cd510f78114c
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>