[FIX] whatsapp: traceback when creating dynamic header text template
Steps:
1. Install the WhatsApp module.
2. Create a WhatsApp template with header type=Text
3. Add Template Header Text = hello {{1}}
On saving this record It will raise traceback

Problem:
the method _compute_variable_ids of the WhatsApp template is trying to fetch
the element of set by header_variable[0] which is not supported.
So it is giving Traceback for that.

Solution:
To solve this issue, use a list for header_variable.

Also added a constraint for dynamic header

Also added a unit test case for the dynamic header

task - 3510168

closes odoo/enterprise#48833

X-original-commit: f6811bd15434b805ad0805f695b27147a8623b1a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>