[FIX] web_studio: edit modifier with no literal value
Versions:
---------
- 15.0
- saas-15.2
- 16.0
- saas-16.1
- saas-16.2

Issue:
------
When you try to modify a modifier with visual studio,
we rewrite the modifier.

```py
modifier = "{
	"key": domain
}"
```

The new modifier will be passed to the `literal_eval` method.
A traceback occurs if domain contains a node that cannot be
evaluated (e.g. an `uid` in a right node of a domain).

Solution:
---------
Place the domain in a string so that it is
evaluated as a string and not a list of tuples.

```py
modifier = "{
	"key": "domain"
}"
```

opw-3491821

closes odoo/enterprise#48208

X-original-commit: 0de8c6049e93ce7a3deb96827ccd419ae6899804
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>