[IMP] test_l10n_be_hr_payroll_account: fix test
Previously, allocation name was not computed on create.
Now, it's not a case anylonger.

With this change, create_allocation test was failing as first the name
field was assigned, then holiday_status_id;
Which caused for the name field to be recomputed and assigned to False.
As the field is required, leave_form.save() was raising an error.

task-3389144

closes odoo/enterprise#47651

Related: odoo/odoo#132187
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>