[IMP] sale_renting: avoid losing rental origin of SO
This commit avoids the recomputation of `is_rental_order`, so that the
fields is not recomputed as soon as the model is edited in a view. In
the context of this task, it allows us to use the catalog in a rental
order and that the order is still considered as a rental order, with
renting dates.

task-3076912

Part-of: odoo/enterprise#47557