[IMP] account_disallowed_expenses{_fleet}: UX improvements
There are three parts to this commit:

1. Date filter
We set it to 'last_year' by default.

2. New warning
A warning is added to the report if the following conditions are satisfied:
- Posted move lines on an expense account within the report dates.
- A vehicle is set on those lines.
- No disallowed expense category has been defined on the account.

This warning contains an action redirecting the user to a list view of these accounts.

3. 'vehicle_split' filter
The filter will now be activated by default when `account_disallowed_expenses` and `fleet` are installed.
This helps the user better understand the report.
Indeed, there could be posted move lines on accounts with a disallowed expense category with AND without a vehicle.
And those lines would otherwise be aggregated and no rate would be displayed.

task-3508394

closes odoo/enterprise#47556

Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>