[IMP] timesheet_grid: Improve tips for list and kanban views
Revise the tips in the timesheet module for the list and Kanban views. The
current information is not applicable in those contexts.

Improve:
Adjust the tip generation based on the view type. This will ensure that tips are
appropriately tailored for list and kanban views, while maintaining the existing
tip for the grid view.

task-3459964

closes odoo/enterprise#45742

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>