[FIX] account_reports: convert fields for analytic that are stored
If there is a column that is stored on `account_move_line`,
but that is not part of the fields `account.move.line`,
it can be a problem.
It could be due to bad migration, or wrong custo by clients.

We will just skip those from the query for the report as they
are indeed useless.

opw-3502834
opw-3499902

closes odoo/enterprise#47872

X-original-commit: 17cc67e314b089271aadd96bf0be1df3523e5a76
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>