[FIX] voip: restore incoming call after fold/unfold
Before this commit: if the dialing panel is folded while a call is
incoming, unfolding it afterwards would cause the call accept/reject
buttons to disappear, making it impossible to do anything but to wait
for the caller to hang up.

After this commit: folding then unfolding the panel doesn't break the
state.

opw-3090255

closes odoo/enterprise#47870

X-original-commit: 1bb69232e7ebbce79a7263a1f9248e665abd996c
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>