[FIX] web_studio: handle case when we get false in action_data
When the user removes module name and '.' (dot seperator) from
template name of a report and then tries to click on 'Qweb Views'
smart button the error will be generated.

To reproduce the issue:
- Turn on developers mode.
- Install studio module.
- Go to Settings > Technical > Actions > Reports.
- Open any report and remove module name and '.' (dot seperator)
from template name of the report.
- Click on 'Qweb Views' smart button the error will be generated.

Error: TypeError: 'bool' object is not subscriptable

The issue is occuring because False is returned from 'associated_view'
method over here -
https://github.com/odoo/odoo/blob/5d98384e446fe8d62679d65baf1caf60bd92611b/odoo/addons/base/models/ir_actions_report.py#L169-L170

and we get above False value in action_data which is not subscriptable -
https://github.com/odoo/enterprise/blob/ee406077a837c208c3b937152b3d5ca41599d85a/web_studio/models/ir_actions_report.py#L54-L55

So now a condition has been added which will check action_data is
False or not and if we get False in action_data we will return the same.

sentry-4484698044

closes odoo/enterprise#48646

X-original-commit: 17fdad481ff30a871e05103dccd209ac9d8b9826
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Saurabh Choraria (sauc) <sauc@odoo.com>