[FIX] mrp_workorder: Field must always be used with Record
Field components must always be used with a record created by the
RelationalModel.

In DocumentViewer, we already have all the necessary data. So we don't
want to use the Composent Record with Field components because we'd have
an extra "web_read". So we will replace the Field components with <img>
and <iframe> which correspond to the rendering of the Field components
in readonly mode.

closes odoo/enterprise#47573

Related: odoo/odoo#135851
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>