[FIX] l10n_lt_reports: Fixes to Balance Sheet
2 bugfixes.

1. The Retained Earnings section of the Balance Sheet should have a
date_scope of `to_beginning_of_fiscalyear`, rather than
`to_beginning_of_period`.
Without this, AMLs between the beginning of the fiscalyear and the
report date will be counted both in Retained Earnings and Current Year
Earnings, leading to an unbalanced Balance Sheet.

2. The Common Summary of Accounts section needs to be added to the total
of Equity.

This is part of the wider task of rebalancing the Balance Sheets.

taskid:3060790

closes odoo/enterprise#47755

X-original-commit: b7a179a3e28a356b45a8fb86eae7f367b3c91b9c
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>