[ADD] knowledge: add article body history
Create a new model mixins that allow other model to
activate the history feature on html fields.

This history automatically track the changes
on the related html field and allow the user
to revert to a previous version at any time.

This also introduce a new component that can be used to :
* List the recent history
* See a previous version of the document
* Compare a previous version to the current one
* Revert the document to a previous version of it

Add and make use of this component in the knowledge topbar

task-3039787

closes odoo/enterprise#37211

Related: odoo/odoo#112957
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>