[FIX] account_reports: journal report double additional_col_x
The function '_get_move_line_additional_col' adds the additional_col_1
and additional_col_2.

In '_get_first_move_line' and '_get_aml_line' when we loop through the
columns in the options, we have to break the loop when we arrive to the
additional columns otherwise they are added twice.

closes odoo/enterprise#47645

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>