[FIX] web_studio: restrict name search for selection pseudo fields
The search functionality in the ExistingFields component allows the enduser to
search for fields present on the model but not yet in the view. However, the
res.users view uses sel_group_id_list pseudo fields indicating membership of
certain security groups. These pseudo fields also have a textual description
that can be searched but no field name as regular fields do.

Before this commit, this caused a traceback when using the search functionality
in dev mode in the presence of such fields: toLowerCase was called on the field
undefined field name. The problem is easily mitigated by checking for a name.

opw-3482276

closes odoo/enterprise#47787

X-original-commit: 170749bd56c222bacc08ff066c758ceed8120f3c
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Tom De Caluwé (tdc) <tdc@odoo.com>