[IMP] sale_renting: quotation templates with rental products
Before this commit, rental products added with a quotation template were
added as 'sold' on a Sales Order, due to the complexity in Rental. Since
the restriction of a single rental period per SO, the inclusion of
rental products is easier.

After this commit, a rental product added to a quotation template
through the Rental App will be 'rented' on the Sales Order.

task-3076912

closes odoo/enterprise#47557

Related: odoo/odoo#135811
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>