[IMP] l10n_ke_hr_payroll: upgrade and new rules
This commit is the continuation of the commit
e43059e73cf7e2374cb80fa90a44fec0c6913dc3.

This commit upgrade the work previously done to master, correct
some typos and refactor the nhif and tax breakdown rules.
This commit also contain some of the adaptation to the new
report mixin.

In addition to that, this commit update some rules and add new ones
to better suit the kenyan current taxes and add required fields to
the payslip.

task-3280883

closes odoo/enterprise#48597

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>