[ADD] pos_order_customer_screen
In this Pr we introduce a new module that enables restaurants to show
their customers a display with all the current orders, filtered by
their preparation state.

In order to know the preparation state of the orders, we rely on the
information provided by the `pos_preparation_display` module.

closes odoo/enterprise#47668

Task: 3514516
Related: odoo/odoo#136065
Signed-off-by: David Monnom (moda) <moda@odoo.com>