[IMP] *: fine-tune app icons
The aim of this commit is to improve the impact and rendering of app
icons in bright and dark mode. It also reduces the size of svg files.

To achieve that, this commit updates the colors to flat colors. This
change will make the icons stand out and improve their readability.

task-3072562

closes odoo/enterprise#48734

X-original-commit: 09bb2d3aed1bb3803144ff0667ab74dc2c0b67e1
Related: odoo/odoo#138279
Signed-off-by: Pierre Paridans (app) <app@odoo.com>