[FIX] sign: sign template not deleting sign item
https://github.com/odoo/enterprise/pull/43103 refactored the sign template
edition page to use the common PDF manipulation logic that all other iframe
components used. However, during this refactoring, the deleteSignItem overwrite
from SignTemplateIframe was calling super before adding the id of the deleted
sign item to the array of deleted sign items. This caused the code to save the
template before saving the id of the deleted item. This commit changes the
order of the super call and the array assignment to fix the behavior.

task-3489667

closes odoo/enterprise#48642

X-original-commit: 00a33fa6672261dd04c3eeb6774157f3f8a688d2
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Leonardo Pavan Rocha <lpr@odoo.com>