[FIX] account_accountant: recompute reco proposition on date change
When trying to reconcile a line with manual
operation with a tax, if we select the
'tax included' option before setting the date,
the tax line isn't recomputed, therefore
a move is created at the reconciliation day date
to reconcile the tax line that has not been
reconciled on the choosen date.
This is because the date change doesn't trigger
the compute.

Steps:

- Create and confirm an invoice in the past
- Reconcile the receivable line with a
  manual operation, select a tax, select
  'tax included in price' and change the date
  in the past
- Validate
-> Two entries are created, one on the reconciliation
   date, an other on today's date for the tax line.

opw-3497532

closes odoo/enterprise#47736

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>