[IMP] mrp_workorder: track time as users employee from backend
When using the shop floor to log in as different employees and then
tracking time on a wo using the buttons in the backend, the employee
that is currently logged in from the shop floor is assigned to this log.
To avoid confusion, tracked time in the backend is now always assigned
to the employee of the currently logged-in users' employee.

Task: 3456827 (nr 8)
Part-of: odoo/enterprise#48379