[IMP] mrp_workorder: continue consumption ordering
When using 'continue consumption' in the shop floor, the newly created
quality check would be shown at the botton of the WO card, instead of
the correct place in the quality check chain.
This commit makes sure to order the quality checks correctly and ensures
that the newly generated register component step is also opened.

Task: 3456827 (nr 12)
Part-of: odoo/enterprise#48379