[IMP] mrp_workorder: default WO time log on assigned employee
When marking a workorder as done which does not have any time logged on
it, a default time log with the expected duration is created
automatically. This commit ensures that if an employee is assigned to
the WO, this employee is also put on this default time log.

Task: 3456827 (nr 19)
Part-of: odoo/enterprise#48379