[IMP] appointment: remove MockServer patch interfering with calendar tests
The appointment module currently extends the CalendarAttendeeController from
the calendar module. Since appointment adds some rpc calls in the controller,
it was decided to patch the MockServer with the appointment routes to prevent
having tests fail because of unimplemented routes. However, patching the
MockServer is not advised by the JS team and the current patch calls ORM mock
methods (like mockRead, mockSearchRead...) that depend on serverData that is
solely defined on the appointment tests. This causes any test done in the
calendar module using the calendar_attendee js class to fail when all modules
are installed because, even though the route is mocked, it tries to fetch data
that is only defined for appointment. This commit changes the approach to use
the mock_server registry, defining each route individually. For routes that are
called on the onWillStart hook and their mock should be doing a mockRead, it
was decided to overwrite the mock inside the appointment tests. This way, it is
possible to return an "empty" object for all tests that are not appointment
related, and for appointment tests we simply overwrite the mock_server for that
route with the original implementation and after the tests are done, we can
restore the mock.

closes odoo/enterprise#47783

Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>