[IMP] knowledge: add collaborative non-editable Ref hook
Generalize the "undo manager" for collaborative controlled content in a
Knowledge Behavior block.

Originally developed for `ArticlesStructureBehavior`, this feature allows an
element to be:
- non-editable (its content is determined programmatically in the Behavior
  Component)
- collaborative (changes of content are shared with collaborators during
  edition)
- restricted to a specific format (there can never be multiple children that
  represent different states of the element as a result of multiple
  collaborators using the `undo` feature of the editor).

The `FileBehavior` will make use of those features for the renaming of the file,
so the logic is exported in a reusable util.

task-3172056

Part-of: odoo/enterprise#43488