[FIX] pos_iot: Use of Worldline in complement of Ingenico
Before commit:
The client is able to use Worldline payment terminal if Ingenico payment terminal is not active in the PoS settings.
If Ingenico payment terminal is activated the worldine payment terminal become invisible to the user in the payment method.

After commit:
The client is able to choose the payment method between Worldline and Ingenico if both of them are activated in the PoS settings.

OPW-3501334

closes odoo/enterprise#48006

X-original-commit: e822b4e4e6853463419aea9b6cb92186408356d4
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Axel Terrier (axte) <axte@odoo.com>