[IMP] account_reports: easier payment registration for bills
Description of the issue/feature this commit addresses:

Currently, the payment registration for vendor bills is kinda hard to figure.
Noticing which bills are overdue, which require a payment is not made so easy
by the interface and could be improved. Also there is no way to register a
payment for multiple lines at a time which could improve the UX.

Desired behavior after the commit is merged :

This commit makes it now possible when in the Journal Items to select
multiple lines and register a payment for them which would create one
payment per line selected.

When inside the Aged Payable report, all amounts are clickables and send the
user to the Journal Items with the proper filter to see what lines are part
of that amount.

When in the Journal Items, the "Create Asset" option was available in
two places. It has been removed from the drop-down "Actions" button and kept
in the header.

task-3516496

closes odoo/enterprise#48679

Related: odoo/odoo#138188
Signed-off-by: William André (wan) <wan@odoo.com>