[IMP] web_studio: adapt code to evalContext clarification
PR odoo/odoo#135782 explicitly adds the "context" key into the
evalContext, whereas before the whole evalContext was considered as
being the context. Studio relied on that loophole to handle the
edition of x2many views, by adding a key "parent" in the context,
which could thus be accessed in expressions without doing
context.get("..."). This commit refactors the way Studio makes the
created datapoint aware of their parent record, without relying on
the context.

Moreover, with odoo/odoo#135782, datapoints makes the assumption
that there's a context in their config, which wasn't the case for
the fake group created in Studio. We thus had to set it.

closes odoo/enterprise#47522

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>