[IMP] account_followup: followup report: clean up, editable title
The main goal of this commit was to make the title of the
followup report editable (and to make the report a bit more standard in
the process).

There were the following technical challenges in making the report title editable.
The render values are not passed to the report like for other standard
qweb reports (e.g. with _get_report_values or computed in the template).
To generate the html of the report (except the external layout including address etc)
the report template just uses a function (res.partner.get_followup_html).
In studio thus only this function call is visible.

The actual report generation basically happens in get_followup_report_html (from account_followup_report.py).
There the render values are computed and the actual report template (template_followup_report) is rendered with these values.

To make the report title editable the call to get_followup_html is removed.
To do this the the necessary values to render the report templates are
obtained directly in the report template (i.e. template_followup_report);
instead of taking the indirection via get_followup_report_html.

task-3446724

Part-of: odoo/enterprise#47633