[FIX] account_reports: fix dispatch_report_action
Currently, the calls are dispatched to the section itself, instead of the main report's custom handler. That is because the `report_id` in the options matches the report the dispatch method is called from.
We should first check `on_sections_source` and call the report with the `sections_source_id` (the source report) with the correct options. Otherwise, the custom handler's actions are not being called, like custom file export.

closes odoo/enterprise#47866

X-original-commit: 73634a4f3d6976e17c634e98ad1ff61725ce9f57
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Aliya Tastemirova (alta) <alta@odoo.com>