[IMP] account_*: separate modes for deferred expenses and revenues
*account_accountant,account_reports

Currently one deferral mode can be selected for both deferred expenses
and deferred revenues. Either the deferrals are created on validation,
when the invoice/bill is posted (perfect when you don't have a lot of
deferrals per month), or a grouped deferral entry is created
manually by the user periodically (interesting when you have a lot of
deferrals per month).

Realistically the situation of a company is not the same concerning
customer invoices and vendor bills. They might generate lots of
customer invoices with deferrals, wanting a manual grouped entry. But
on the other hand they might not have a lot of vendor bills with
deferrals, so don't want to create a manual entry for this every month.

In order to facilitate this, we now allow to select a different mode
for both deferred revenues and deferred expenses.

task-3516660

Part-of: odoo/enterprise#48022